Investigation, planning
I want to premap regions for the text section of the running binary for each process and the stack (working with single-thread for now). In the restorer, I will call remap to relocate those regions instead of mapping them outright. For the stack, it will be a bit more complicated than in the binary since it looks like there will already be an io vector rendered for the stack (reading from the pages image), so I will have to read from the pages image before the restorer does.